Correct name of application main class
When trying to build the doran_and_parberry application, one receives the
following error:

Export/html5/haxe/ApplicationMain.hx:267: characters 7-42 : Class not found : Main

Checking the file `ApplicationMain.hx`, one sees that the class is actually
called `ApplicationMain`.  This change to the project file corrects the
error and allows the application to be built.